chore: move utils to /internal
